moved PassThru tests to a separate module; added some param action regions; added perfunctory action str/repr test
